requirements.txt
pystyle
pyyaml
colorama
